chore: fix clippy lints under -D warnings on host / Android / iOS targets
Pre-existing lints surfaced when running
`cargo clippy --workspace --all-targets -- -D warnings` against the
host workspace, then again separately against the iOS and Android
cross-compile targets. None of these are related to specific bug
fixes — they're all CI hardening so the next time clippy runs in CI
with `-D warnings` we don't break the build.

  * `blinc_theme::platform::mod` — removed `return` statements from
    cfg-gated arms in `detect_system_color_scheme` so each compiled
    branch returns its expression directly.

  * `blinc_layout::motion` / `overlay_state` and `blinc_router` —
    added module-level
    `#![allow(clippy::missing_const_for_thread_local)]`. Nightly
    clippy 0.1.96+ mis-fires this lint on `thread_local! { static X = const { ... } }`
    even though the initializer IS already wrapped in `const { ... }`.
    Suppressed at the module level because the lint attaches to the
    macro expansion, not the static item.

  * `blinc_platform_android::{activity, event_loop}` — replaced
    single-arm `match` with `if let` per `clippy::single_match`.

  * `blinc_platform_ios::event_loop` — added missing `Default` impls
    for `IOSWakeProxy` and `IOSEventLoop`, dropped unused `debug` /
    `warn` imports, renamed unused parameter `handler` →
    `_handler`.

  * `blinc_platform_ios::app` — `UIScreen::mainScreen` is deprecated
    on modern iOS, but at platform-init time we don't yet have a
    view / window / scene to use the recommended replacement against;
    explicitly `#[allow(deprecated)]` with the rationale. Also
    removed an `as f64` cast that was casting an `f64` to `f64`.

  * `blinc_platform_ios::assets` — removed unnecessary `unsafe`
    block around `NSBundle::mainBundle()` (no unsafe ops inside).

  * `blinc_platform_ios::lib` — moved `use blinc_platform::PlatformError`
    behind `#[cfg(not(target_os = "ios"))]` since it's only used in
    the non-iOS placeholder impl.

After this commit, all three of these are clean:

  cargo clippy --workspace --all-targets -- -D warnings
  cargo clippy -p blinc_app --target aarch64-linux-android \
      --features android --no-default-features -- -D warnings
  cargo clippy -p blinc_app --target aarch64-apple-ios \
      --features ios --no-default-features -- -D warnings

+++ extensions/blinc_platform_ios/src/app.rs
@@ -27,11 +27,21 @@ pub struct IOSPlatform {
 impl IOSPlatform {
     /// Get the main screen's scale factor
     fn get_screen_scale() -> f64 {
-        // On iOS, UIScreen::mainScreen() requires MainThreadMarker
-        // We assume this is called from the main thread
+        // On iOS, UIScreen::mainScreen() requires MainThreadMarker.
+        // We assume this is called from the main thread.
+        //
+        // `mainScreen` is technically deprecated in modern iOS (Apple
+        // recommends `view.window.windowScene.screen` instead) but at
+        // platform-init time we don't have a view / window / scene to
+        // hand — those don't exist until the UIApplicationDelegate
+        // boots up. The deprecated form continues to work for
+        // single-screen apps, which is the only configuration Blinc's
+        // iOS runner currently supports.
+        #[allow(deprecated)]
         let mtm = MainThreadMarker::new().expect("Must be called from main thread");
+        #[allow(deprecated)]
         let screen = UIScreen::mainScreen(mtm);
-        screen.scale() as f64
+        screen.scale()
     }
 }
